refactor(caps): split capability seam into focused modules
The monolithic caps module was split into a thin export-focused facade plus a new ops module and the existing sibling adapter modules, moving capability construction, curation, preflight, and invocation logic out of the single file. This is a pure reorganization with no behavior change; all existing `caps::X` paths continue to resolve through the preserved glob re-exports.
